Switch base image to node:alpine
Alpine is a lightweight Linux distribution designed for use in containers and aims to offer the required tools and packages at the minimum size.

The official Alpine-based Node.js image is 67MB, while the default Debian-based one is a whopping 672MB. Both images provide `yarn` by default, we just need to add `git` to be able to install `@hollowverse/common`.

One caveat is that the OpenSSL version in Alpine (and also in Ubuntu) is the 1.0.2k, which is the LTS version and is still receiving security patches but the encrypted secrets in the other repos are incompatible because they were encrypted with OpenSSL 1.1.

I'd like to remove the compilation commands we are currently doing and re-encrypt the secrets with the LTS version instead.

The build environment image built from this branch is 187MB and takes  2m 36s to build on my laptop. The same image built from master is 1.02GB and takes 17m 19s to build on the same laptop with equivalant cached layers!

This should result in faster builds and deployments.