Fail if yarn.lock is not up-to-date
With clown updating our `package.json` files, it's easy to forget to run `yarn` after `package.json` is updated. This means that `yarn.lock` could go out of sync with `package.json` so we might get some dependencies that are not locked. `--frozen-lockfile` fails the build if that is the case.